Organise storage portlayer consistently
This brings consistent organisation and naming of files to the storage
portlayer in preparation for moving Image to use the standard Join/Bind
semantic.

There are some additional structure naming changes that could be made for
consistency, and some leaking abstractions (e.g. exposing volume type to
handlers) that should be addressed
