+++ lib/apiservers/portlayer/restapi/handlers/storage_handlers.go
@@ -31,7 +31,12 @@ import (
 	"github.com/vmware/vic/lib/archive"
 	epl "github.com/vmware/vic/lib/portlayer/exec"
 	spl "github.com/vmware/vic/lib/portlayer/storage"
-	"github.com/vmware/vic/lib/portlayer/storage/nfs"
+	"github.com/vmware/vic/lib/portlayer/storage/container"
+	"github.com/vmware/vic/lib/portlayer/storage/image"
+	vsimage "github.com/vmware/vic/lib/portlayer/storage/image/vsphere"
+	"github.com/vmware/vic/lib/portlayer/storage/volume"
+	"github.com/vmware/vic/lib/portlayer/storage/volume/nfs"
+	vsvolume "github.com/vmware/vic/lib/portlayer/storage/volume/vsphere"
 	"github.com/vmware/vic/lib/portlayer/storage/vsphere"
 	"github.com/vmware/vic/lib/portlayer/util"
 	"github.com/vmware/vic/pkg/trace"
@@ -40,9 +45,9 @@ import (
 
 // StorageHandlersImpl is the receiver for all of the storage handler methods
 type StorageHandlersImpl struct {
-	imageCache     *spl.NameLookupCache
-	volumeCache    *spl.VolumeLookupCache
-	containerStore *vsphere.ContainerStore
+	imageCache     *image.NameLookupCache
+	volumeCache    *volume.VolumeLookupCache
+	containerStore *container.ContainerStore
 }
 
 const (
@@ -69,7 +74,7 @@ func (h *StorageHandlersImpl) Configure(api *operations.PortLayerAPI, handlerCtx
 		op.Warnf("Multiple image stores found. Multiple image stores are not yet supported. Using [%s] %s", imageStoreURL.Host, imageStoreURL.Path)
 	}
 
-	ds, err := vsphere.NewImageStore(op, handlerCtx.Session, &imageStoreURL)
+	ds, err := vsimage.NewImageStore(op, handlerCtx.Session, &imageStoreURL)
 	if err != nil {
 		op.Panicf("Cannot instantiate storage layer: %s", err)
 	}
@@ -77,12 +82,12 @@ func (h *StorageHandlersImpl) Configure(api *operations.PortLayerAPI, handlerCtx
 	// The imagestore is implemented via a cache which is backed via an
 	// implementation that writes to disks.  The cache is used to avoid
 	// expensive metadata lookups.
-	h.imageCache = spl.NewLookupCache(ds)
+	h.imageCache = image.NewLookupCache(ds)
 
 	spl.RegisterImporter(op, imageStoreURL.String(), ds)
 	spl.RegisterExporter(op, imageStoreURL.String(), ds)
 
-	c, err := vsphere.NewContainerStore(op, handlerCtx.Session, h.imageCache)
+	c, err := container.NewContainerStore(op, handlerCtx.Session, h.imageCache)
 	if err != nil {
 		op.Panicf("Couldn't create containerStore: %s", err.Error())
 	}
@@ -114,11 +119,11 @@ func (h *StorageHandlersImpl) Configure(api *operations.PortLayerAPI, handlerCtx
 
 func (h *StorageHandlersImpl) configureVolumeStores(op trace.Operation, handlerCtx *HandlerContext) {
 	var (
-		vs  spl.VolumeStorer
+		vs  volume.VolumeStorer
 		err error
 	)
 
-	h.volumeCache = spl.NewVolumeLookupCache(op)
+	h.volumeCache = volume.NewVolumeLookupCache(op)
 
 	// register the pseudo-store to handle the generic "volume" store name
 	spl.RegisterImporter(op, "volume", h.volumeCache)
@@ -248,7 +253,7 @@ func (h *StorageHandlersImpl) DeleteImage(params storage.DeleteImageParams) midd
 		return ferr(err, http.StatusInternalServerError)
 	}
 
-	image, err := spl.Parse(imageURL)
+	img, err := image.Parse(imageURL)
 	if err != nil {
 		return ferr(err, http.StatusInternalServerError)
 	}
@@ -263,11 +268,11 @@ func (h *StorageHandlersImpl) DeleteImage(params storage.DeleteImageParams) midd
 		keepNodes[idx] = k
 	}
 
-	op := trace.NewOperation(context.Background(), fmt.Sprintf("DeleteBranch(%s)", image.ID))
-	deletedImages, err := h.imageCache.DeleteBranch(op, image, keepNodes)
+	op := trace.NewOperation(context.Background(), fmt.Sprintf("DeleteBranch(%s)", img.ID))
+	deletedImages, err := h.imageCache.DeleteBranch(op, img, keepNodes)
 	if err != nil {
 		switch {
-		case spl.IsErrImageInUse(err):
+		case image.IsErrImageInUse(err):
 			return ferr(err, http.StatusLocked)
 
 		case os.IsNotExist(err):
@@ -279,8 +284,8 @@ func (h *StorageHandlersImpl) DeleteImage(params storage.DeleteImageParams) midd
 	}
 
 	result := make([]*models.Image, len(deletedImages))
-	for idx, image := range deletedImages {
-		result[idx] = convertImage(image)
+	for idx, img := range deletedImages {
+		result[idx] = convertImage(img)
 	}
 
 	return storage.NewDeleteImageOK().WithPayload(result)
@@ -326,7 +331,7 @@ func (h *StorageHandlersImpl) WriteImage(params storage.WriteImageParams) middle
 			})
 	}
 
-	parent := &spl.Image{
+	parent := &image.Image{
 		Store: u,
 		ID:    params.ParentID,
 	}
@@ -396,7 +401,7 @@ func (h *StorageHandlersImpl) CreateVolume(params storage.CreateVolumeParams) mi
 	}
 
 	op := trace.NewOperation(context.Background(), fmt.Sprintf("VolumeCreate(%s)", params.VolumeRequest.Name))
-	volume, err := h.volumeCache.VolumeCreate(op, params.VolumeRequest.Name, storeURL, capacity*1024, byteMap)
+	vol, err := h.volumeCache.VolumeCreate(op, params.VolumeRequest.Name, storeURL, capacity*1024, byteMap)
 	if err != nil {
 
 		if os.IsExist(err) {
@@ -408,7 +413,7 @@ func (h *StorageHandlersImpl) CreateVolume(params storage.CreateVolumeParams) mi
 		}
 
 		op.Errorf("storagehandler: VolumeCreate error: %#v", err)
-		if _, ok := err.(spl.VolumeStoreNotFoundError); ok {
+		if _, ok := err.(volume.VolumeStoreNotFoundError); ok {
 			return storage.NewCreateVolumeNotFound().WithPayload(&models.Error{
 				Code:    http.StatusNotFound,
 				Message: err.Error(),
@@ -421,7 +426,7 @@ func (h *StorageHandlersImpl) CreateVolume(params storage.CreateVolumeParams) mi
 		})
 	}
 
-	response := volumeToCreateResponse(volume, params.VolumeRequest)
+	response := volumeToCreateResponse(vol, params.VolumeRequest)
 	return storage.NewCreateVolumeCreated().WithPayload(&response)
 }
 
@@ -463,7 +468,7 @@ func (h *StorageHandlersImpl) RemoveVolume(params storage.RemoveVolumeParams) mi
 				Message: err.Error(),
 			})
 
-		case spl.IsErrVolumeInUse(err):
+		case volume.IsErrVolumeInUse(err):
 			return storage.NewRemoveVolumeConflict().WithPayload(&models.Error{
 				Message: err.Error(),
 			})
@@ -530,11 +535,13 @@ func (h *StorageHandlersImpl) VolumeJoin(params storage.VolumeJoinParams) middle
 		})
 	}
 
+	// NOTE: unclear to me why we are leaking this logic at this level - the volume should be able to switch Join implementations
+	// based on its type
 	switch volume.Device.DiskPath().Scheme {
 	case nfsScheme:
 		actualHandle, err = nfs.VolumeJoin(op, actualHandle, volume, params.JoinArgs.MountPath, params.JoinArgs.Flags)
 	case dsScheme:
-		actualHandle, err = vsphere.VolumeJoin(op, actualHandle, volume, params.JoinArgs.MountPath, params.JoinArgs.Flags)
+		actualHandle, err = vsvolume.VolumeJoin(op, actualHandle, volume, params.JoinArgs.MountPath, params.JoinArgs.Flags)
 	default:
 		err = fmt.Errorf("unknown scheme (%s) for Volume (%#v)", volume.Device.DiskPath().Scheme, *volume)
 	}
@@ -683,7 +690,7 @@ func (h *StorageHandlersImpl) StatPath(params storage.StatPathParams) middleware
 //utility functions
 
 // convert an SPL Image to a swagger-defined Image
-func convertImage(image *spl.Image) *models.Image {
+func convertImage(image *image.Image) *models.Image {
 	var parent, selfLink string
 
 	// scratch image
@@ -711,7 +718,7 @@ func convertImage(image *spl.Image) *models.Image {
 	}
 }
 
-func volumeToCreateResponse(volume *spl.Volume, model *models.VolumeRequest) models.VolumeResponse {
+func volumeToCreateResponse(volume *volume.Volume, model *models.VolumeRequest) models.VolumeResponse {
 	response := models.VolumeResponse{
 		Driver:   model.Driver,
 		Name:     volume.ID,
@@ -722,7 +729,7 @@ func volumeToCreateResponse(volume *spl.Volume, model *models.VolumeRequest) mod
 	return response
 }
 
-func fillVolumeModel(volume *spl.Volume) (models.VolumeResponse, error) {
+func fillVolumeModel(volume *volume.Volume) (models.VolumeResponse, error) {
 	storeName, err := util.VolumeStoreName(volume.Store)
 	if err != nil {
 		return models.VolumeResponse{}, err
@@ -741,7 +748,7 @@ func fillVolumeModel(volume *spl.Volume) (models.VolumeResponse, error) {
 	return model, nil
 }
 
-func createMetadataMap(volume *spl.Volume) map[string]string {
+func createMetadataMap(volume *volume.Volume) map[string]string {
 	stringMap := make(map[string]string)
 	for k, v := range volume.Info {
 		stringMap[k] = string(v)
@@ -749,7 +756,7 @@ func createMetadataMap(volume *spl.Volume) map[string]string {
 	return stringMap
 }
 
-func createNFSVolumeStore(op trace.Operation, dsurl *url.URL, name string) (spl.VolumeStorer, error) {
+func createNFSVolumeStore(op trace.Operation, dsurl *url.URL, name string) (volume.VolumeStorer, error) {
 	var err error
 	uid, gid, err := parseUIDAndGID(dsurl)
 	if err != nil {
@@ -805,7 +812,7 @@ func parseUIDAndGID(queryURL *url.URL) (int, int, error) {
 	return uid, gid, nil
 }
 
-func createVsphereVolumeStore(op trace.Operation, dsurl *url.URL, name string, handlerCtx *HandlerContext) (spl.VolumeStorer, error) {
+func createVsphereVolumeStore(op trace.Operation, dsurl *url.URL, name string, handlerCtx *HandlerContext) (volume.VolumeStorer, error) {
 	ds, err := datastore.NewHelperFromURL(op, handlerCtx.Session, dsurl)
 	if err != nil {
 		err = fmt.Errorf("cannot find datastores: %s", err)
@@ -813,7 +820,7 @@ func createVsphereVolumeStore(op trace.Operation, dsurl *url.URL, name string, h
 		return nil, err
 	}
 
-	vs, err := vsphere.NewVolumeStore(op, name, handlerCtx.Session, ds)
+	vs, err := vsvolume.NewVolumeStore(op, name, handlerCtx.Session, ds)
 	if err != nil {
 		err = fmt.Errorf("cannot instantiate the volume store: %s", err)
 		op.Errorf("%s", err.Error())
