+++ lib/portlayer/storage/volume/nfs/store.go
@@ -27,6 +27,7 @@ import (
 	"github.com/vmware/vic/lib/archive"
 	"github.com/vmware/vic/lib/config/executor"
 	"github.com/vmware/vic/lib/portlayer/storage"
+	"github.com/vmware/vic/lib/portlayer/storage/volume"
 	"github.com/vmware/vic/lib/portlayer/util"
 	"github.com/vmware/vic/pkg/trace"
 	"github.com/vmware/vic/pkg/vsphere/vm"
@@ -114,7 +115,7 @@ func (v *VolumeStore) volMetadataDirPath(ID string) string {
 }
 
 // Creates a volume directory and volume object for NFS based volumes
-func (v *VolumeStore) VolumeCreate(op trace.Operation, ID string, store *url.URL, capacityKB uint64, info map[string][]byte) (*storage.Volume, error) {
+func (v *VolumeStore) VolumeCreate(op trace.Operation, ID string, store *url.URL, capacityKB uint64, info map[string][]byte) (*volume.Volume, error) {
 	target, err := v.Service.Mount(op)
 	if err != nil {
 		return nil, err
@@ -134,7 +135,7 @@ func (v *VolumeStore) VolumeCreate(op trace.Operation, ID string, store *url.URL
 		return nil, fmt.Errorf("Unexpected scheme (%s) for volume store (%s)", u.Scheme, v.Name)
 	}
 
-	vol, err := storage.NewVolume(v.SelfLink, ID, info, NewVolume(u, v.volDirPath(ID)), executor.CopyNew)
+	vol, err := volume.NewVolume(v.SelfLink, ID, info, NewVolume(u, v.volDirPath(ID)), executor.CopyNew)
 	if err != nil {
 		op.Errorf("Created volume directory but failed to create volume: %s", err)
 		return nil, err
@@ -149,7 +150,7 @@ func (v *VolumeStore) VolumeCreate(op trace.Operation, ID string, store *url.URL
 }
 
 // VolumeDestroy Removes a volume and all of its contents from the nfs store. We already know via the cache if it is in use.
-func (v *VolumeStore) VolumeDestroy(op trace.Operation, vol *storage.Volume) error {
+func (v *VolumeStore) VolumeDestroy(op trace.Operation, vol *volume.Volume) error {
 	target, err := v.Service.Mount(op)
 	if err != nil {
 		return err
@@ -173,7 +174,7 @@ func (v *VolumeStore) VolumeDestroy(op trace.Operation, vol *storage.Volume) err
 	return nil
 }
 
-func (v *VolumeStore) VolumesList(op trace.Operation) ([]*storage.Volume, error) {
+func (v *VolumeStore) VolumesList(op trace.Operation) ([]*volume.Volume, error) {
 
 	target, err := v.Service.Mount(op)
 	if err != nil {
@@ -185,7 +186,7 @@ func (v *VolumeStore) VolumesList(op trace.Operation) ([]*storage.Volume, error)
 	if err != nil {
 		return nil, err
 	}
-	var volumes []*storage.Volume
+	var volumes []*volume.Volume
 	var fetchErr error
 
 	for _, fileInfo := range volFileInfo {
@@ -203,7 +204,7 @@ func (v *VolumeStore) VolumesList(op trace.Operation) ([]*storage.Volume, error)
 
 		// #nosec: Errors unhandled.
 		u, _ := v.Service.URL()
-		vol, err := storage.NewVolume(v.SelfLink, fileInfo.Name(), volMetadata, NewVolume(u, v.volDirPath(fileInfo.Name())), executor.CopyNew)
+		vol, err := volume.NewVolume(v.SelfLink, fileInfo.Name(), volMetadata, NewVolume(u, v.volDirPath(fileInfo.Name())), executor.CopyNew)
 		if err != nil {
 			op.Errorf("Failed to create volume struct from volume directory (%s)", fileInfo.Name())
 			return nil, err
