+++ lib/portlayer/storage/volume/vsphere/export.go
@@ -24,6 +24,8 @@ import (
 	"github.com/vmware/vic/lib/archive"
 	"github.com/vmware/vic/lib/guest"
 	"github.com/vmware/vic/lib/portlayer/storage"
+	"github.com/vmware/vic/lib/portlayer/storage/volume"
+	"github.com/vmware/vic/lib/portlayer/storage/vsphere"
 	"github.com/vmware/vic/pkg/trace"
 	"github.com/vmware/vic/pkg/vsphere/disk"
 	"github.com/vmware/vic/pkg/vsphere/vm"
@@ -126,94 +128,9 @@ func (v *VolumeStore) newDataSource(op trace.Operation, url *url.URL) (storage.D
 func (v *VolumeStore) newOnlineDataSource(op trace.Operation, owner *vm.VirtualMachine, id string) (storage.DataSource, error) {
 	op.Debugf("Constructing toolbox data source: %s.%s", owner.Reference(), id)
 
-	return &ToolboxDataSource{
+	return &vsphere.ToolboxDataSource{
 		VM:    owner,
-		ID:    storage.Label(id),
+		ID:    volume.Label(id),
 		Clean: func() { return },
 	}, nil
-}
-
-// Export reads the delta between child and parent image layers, returning
-// the difference as a tar archive.
-//
-// id - must inherit from ancestor if ancestor is specified
-// ancestor - the layer up the chain against which to diff
-// spec - describes filters on paths found in the data (include, exclude, rebase, strip)
-// data - set to true to include file data in the tar archive, false to include headers only
-func (i *ImageStore) Export(op trace.Operation, id, ancestor string, spec *archive.FilterSpec, data bool) (io.ReadCloser, error) {
-	l, err := i.NewDataSource(op, id)
-	if err != nil {
-		return nil, err
-	}
-
-	if ancestor == "" {
-		return l.Export(op, spec, data)
-	}
-
-	// for now we assume ancestor instead of entirely generic left/right
-	// this allows us to assume it's an image
-	r, err := i.NewDataSource(op, ancestor)
-	if err != nil {
-		op.Debugf("Unable to get datasource for ancestor: %s", err)
-
-		l.Close()
-		return nil, err
-	}
-
-	closers := func() error {
-		op.Debugf("Callback to io.Closer function for image delta export")
-
-		l.Close()
-		r.Close()
-
-		return nil
-	}
-
-	ls := l.Source()
-	rs := r.Source()
-
-	fl, lok := ls.(*os.File)
-	fr, rok := rs.(*os.File)
-
-	if !lok || !rok {
-		go closers()
-		return nil, fmt.Errorf("mismatched datasource types: %T, %T", ls, rs)
-	}
-
-	// if we want data, exclude the xattrs, otherwise assume diff
-	tar, err := archive.Diff(op, fl.Name(), fr.Name(), spec, data, !data)
-	if err != nil {
-		go closers()
-		return nil, err
-	}
-
-	return &storage.ProxyReadCloser{
-		ReadCloser: tar,
-		Closer:     closers,
-	}, nil
-}
-
-func (i *ImageStore) NewDataSource(op trace.Operation, id string) (storage.DataSource, error) {
-	url, err := i.URL(op, id)
-	if err != nil {
-		return nil, err
-	}
-
-	return i.newDataSource(op, url)
-}
-
-func (i *ImageStore) newDataSource(op trace.Operation, url *url.URL) (storage.DataSource, error) {
-	mountPath, cleanFunc, err := i.Mount(op, url, false)
-	if err != nil {
-		return nil, err
-	}
-
-	f, err := os.Open(mountPath)
-	if err != nil {
-		cleanFunc()
-		return nil, err
-	}
-
-	op.Debugf("Created mount data source for access to %s at %s", url, mountPath)
-	return storage.NewMountDataSource(op, f, cleanFunc), nil
 }
